Bug 1525719: remove eyeball and gear menus, move languages and edit to header
This removes server-side user-dependent code from the jinja template
for our /ducks/ URL and simplifies the UX the user sees by removing
controls from wiki page titles and moving some of them to the main
header. The subscribe and advanced menus are removed. The language
menu and the edit button are moved to the header. The edit link is
only visible when the user is logged in. Then language menu is not
yet implemented (because it will need a document API to work with
client-side navigation).

Additionally, this patch also consolidates all of the SVG icon files
in a new icons/ directory instead of having them in the same directory
as the .jsx files.
