+++ kuma/wiki/jinja2/wiki/react_document.html
@@ -111,11 +111,6 @@
       <div class="document-title">
         <h1>{{ document.title }}</h1>
       </div>
-
-      <!-- action buttons -->
-      <div class="document-actions">
-        {{ buttons }}
-      </div>
     </div>
   </div>
   {% if toc_html %}
